Remove `z` value for (partial) alignments
This used to be one of three similar alignment score measures. By removing
it, we require alignment scores to be negative (or zero). Also, the maximal
allowed cumulative penalty has to be negative. This patch also
introduces a helper struct that pleases the linter since the maximum number
of function arguments is not exceeded anymore.
During these changes, the order of the arguments to the assert!-macro was
reversed in order to get more helpful failure messages from the IDE.