Improve Robustness of Distributed Mapping
The dispatcher now memorizes the assignments of tasks to worker
nodes and requeries tasks that were assigned to failed workers.
With this, worker nodes can be spawned and destroyed anytime
during the mapping process which can be useful in distributed
computing environments.
This commit also cleans up older code that aimed at the same
target but was never finished.