Scale MQ Values Down to Reflect Uncertainty
When we found a best-scoring hit, we search up until `AS + mm_penalty` to find suboptimal
hits in order to estimate the mapping quality. However, we don't ever search beyond the
`mismatch_bound` limit. So when we are closer than `mm_penalty` to the `mismatch_bound`,
the mapping quality needs to be scaled down proportionally to the difference between
`mm_bound` and `AS` to reflect the fact that we are less likely to find suboptimal
alignments within these narrowed bounds.